Added performance chart but it still sucks
